old_files/backend/requirements.txt
bardapi==0.1.33
elevenlabs==0.2.24
fastapi==0.102.0
openai==0.27.9
python-dotenv==1.0.0
uvicorn==0.23.2
Requests==2.31.0
httpx==0.24.1
elevenlabs==0.2.24
